Backend/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.1.0
idna==3.3
importlib-metadata==4.11.4
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
mediapipe==0.8.11
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
pywin32==304
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
six==1.15.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.10.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
uptime==3.0.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
